Split ProposePotentialNewUserSteps into OktaEnabled and OktaDisabled feature files.
